Add portable safe_string implementations
This is probably a horrible idea, my original plan was to replace all uses
of _s functions with std::string where possible.

But they are everywhere in the codebase. and we need to do this port in
smaller steps. So this is the more sane solution in the short term.

Ideally we will remove these later